requirements.txt
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==3.0.3
numpy==1.16.2
numpy-stl==2.10.1
pyparsing==2.4.0
python-dateutil==2.8.0
python-utils==2.3.0
six==1.12.0
memoized==0.3